1.update package name 2.support build.gradle and gradle.properties 3.support ProtocolTransportObject
