+++ eventmesh-protocol-plugin/eventmesh-protocol-meshmessage/build.gradle
@@ -15,9 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.protocol.eventmeshmessage;
+dependencies {
+    compileOnly project(":eventmesh-protocol-plugin:eventmesh-protocol-api")
+    implementation "io.cloudevents:cloudevents-core"
 
-public enum EventMeshMessageProtocolConstant {
-    ;
-    public static final String PROTOCOL_NAME = "eventmeshmessage";
+    testImplementation project(":eventmesh-protocol-plugin:eventmesh-protocol-api")
+    testImplementation "io.cloudevents:cloudevents-core"
+    testImplementation "junit:junit"
 }
