+++ eventmesh-protocol-plugin/eventmesh-protocol-meshmessage/gradle.properties
@@ -12,5 +12,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+#
 
-eventmeshmessage=org.apache.eventmesh.protocol.eventmeshmessage.EventMeshMessageProtocolAdaptor
+pluginType=protocol
+pluginName=eventmeshmessage
