+++ settings.gradle
@@ -37,5 +37,5 @@ include 'eventmesh-protocol-plugin'
 include 'eventmesh-protocol-plugin:eventmesh-protocol-api'
 include 'eventmesh-protocol-plugin:eventmesh-protocol-openmessage'
 include 'eventmesh-protocol-plugin:eventmesh-protocol-cloudevents'
-include 'eventmesh-protocol-plugin:eventmesh-protocol-eventmeshmessage'
+include 'eventmesh-protocol-plugin:eventmesh-protocol-meshmessage'
